% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{hp_class}
\alias{hp_class}
\title{HP class (Example data)}
\format{A named list of 10962 character vectors.}
\source{
\url{http://compbio.charite.de/hudson/job/hpo/1529/artifact/hp/hp.obo}
}
\description{
Each HP term can be of one or several classes.
Classes are HP terms direct descendants of the 'Phenotypic abnormality'
term.
}
\details{
These data are used to examplify the different functions of the
package.
More data are available in the MultiHumanPhenoDB package.
}
\examples{
data(hpDef, hp_class, package="PCAN")
hp <- "HP:0100089"
hpDef[which(hpDef$id==hp),]
# This term has 2 classes:
hpDef[which(hpDef$id \%in\% hp_class[[hp]]),] 
}

