% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{hsEntrezByRPath}
\alias{hsEntrezByRPath}
\title{Homo sapiens entrez gene ID by Reactome pathway (Example data)}
\format{A named list of 1345 character vectors.}
\source{
Two ressources were used in June 2 2015: \itemize{
\item{\url{http://www.reactome.org/download/current/UniProt2Reactome.txt}
was used to get list of Uniprot ID associated to each pathway.}
\item{
\url{ftp://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/idmapping/by_organism/HUMAN_9606_idmapping.dat.gz}
was used to map Uniprot ID to Entrez gene IDs}
}
}
\description{
The human genes coding for proteins involved in
the different Reactome pathways.
}
\examples{
\dontrun{example(hpGeneListComp)}
}

