% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_list_density.R
\name{net_list_density}
\alias{net_list_density}
\title{Calculate the mean subgraph density matrix}
\usage{
net_list_density(netlist, directed = FALSE, configuration)
}
\arguments{
\item{netlist}{A list of networks in a form or adjacency matrix}

\item{directed}{Default is FALSE (undirected network).}

\item{configuration}{Collection of network statistics (topological features)}
}
\description{
Calculate the mean subgraph density matrix
}
