% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_list_density_partition.R
\name{net_list_density_partition}
\alias{net_list_density_partition}
\title{Calculate the mean subgraph density matrix after partitioning}
\usage{
net_list_density_partition(netlist, directed = FALSE, configuration,
  tau = NULL, K = NULL, seed = 1234)
}
\arguments{
\item{directed}{Default is FALSE (undirected network).}

\item{configuration}{Collection of network statistics (topological features)}

\item{tau}{Minimum size of a subgraph}

\item{K}{The number of subgraphs}

\item{seed}{Seed number. Default is 1234.}

\item{net}{A network (adjacency matrix)}

\item{minsize}{Minimum size of a network for determining K and tau. Default is the size of the network net.}
}
\description{
Calculate the mean subgraph density matrix after partitioning
}
