% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_stat.R
\name{net_stat}
\alias{net_stat}
\title{Calculate network statistics}
\usage{
net_stat(net, directed = FALSE, configuration = c("edges", "isolate",
  "kstar(2)", "kstar(3)", "kstar(4)", "kstar(5)", "cycle(4)", "cycle(5)",
  "cycle(6)"))
}
\arguments{
\item{net}{Input network (adjacency matrix)}

\item{stat}{Statistics}
}
\description{
Calculate network statistics
}
