% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAN.R
\name{PCAN}
\alias{PCAN}
\title{PCA on a sample of networks and calculate PCA contribution of variables to explaining variabilities in PCs}
\usage{
PCAN(netlist, directed = FALSE, configuration, numdim = 5,
  subgroup = NULL)
}
\arguments{
\item{netlist}{A list of networks in a form or adjacency matrix}

\item{directed}{Default is FALSE (undirected network).}

\item{configuration}{Collection of network statistics (topological features)}

\item{numdim}{Number of PCs for calculating contribution. Default is 5.}

\item{subgroup}{Vector of subgroup membership. Default is NULL (i.e. no subgroup).}

\item{K}{The number of subgraphs}

\item{tau}{Minimum size of a subgraph}

\item{seed}{Seed number. Default is 1234.}
}
\description{
PCA on a sample of networks and calculate PCA contribution of variables to explaining variabilities in PCs
}
