% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{rPath}
\alias{rPath}
\title{Reactome pathways (Example data)}
\format{A data frame with 1345 rows and 2 columns:
\describe{
\item{Pathway}{Reactome ID.}
\item{Pathway_name}{The name of the pathway.}
}}
\source{
\url{http://www.reactome.org/download/current/UniProt2Reactome.txt}
}
\description{
Pathways taken from the Reactome database.
}
\examples{
\dontrun{example(hpGeneListComp)}
}

